refactor: 식당 생성 메서드 내부 메서드 전환 및 유저 정보 회 서비스 이름 변경 ( #49 )
