+++ src/test/java/jpabook/dashdine/service/restaurant/RestaurantManagementServiceTest.java
@@ -5,7 +5,7 @@ import jpabook.dashdine.domain.user.User;
 import jpabook.dashdine.domain.user.UserRoleEnum;
 import jpabook.dashdine.dto.request.restaurant.CreateRestaurantParam;
 import jpabook.dashdine.repository.restaurant.RestaurantRepository;
-import jpabook.dashdine.service.user.UserInfoService;
+import jpabook.dashdine.service.user.query.UserQueryService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -29,7 +29,7 @@ class RestaurantManagementServiceTest {
     private RestaurantRepository restaurantRepository;
 
     @Mock
-    private UserInfoService userInfoService;
+    private UserQueryService userQueryService;
 
     private User user;
 
@@ -49,7 +49,7 @@ class RestaurantManagementServiceTest {
         @DisplayName("생성 실패 : 중복된 이름")
         void failToCreateRestaurantWithDuplicateName() {
             // Given
-            when(userInfoService.findUser(user.getLoginId())).thenReturn(user);
+            when(userQueryService.findUser(user.getLoginId())).thenReturn(user);
             when(restaurantRepository.findRestaurantNameByUserId(user.getId())).thenReturn(asList("exampleName1", "exampleName2"));
 
             // When & Then
@@ -64,7 +64,7 @@ class RestaurantManagementServiceTest {
         @DisplayName("생성 성공 : 중복 이름 없음")
         void successToCreateRestaurant(){
             // Given
-            when(userInfoService.findUser(user.getLoginId())).thenReturn(user);
+            when(userQueryService.findUser(user.getLoginId())).thenReturn(user);
             when(restaurantRepository.findRestaurantNameByUserId(user.getId())).thenReturn(asList("exampleName2", "exampleName3"));
 
             // When
